+++ pkg/storegateway/gateway_ring.go
@@ -34,11 +34,6 @@ const (
 	// a different name to avoid clashing Prometheus metrics when running in single-binary).
 	RingNameForClient = "store-gateway-client"
 
-	// RingNumTokens is the number of tokens registered in the ring by each store-gateway
-	// instance. We use a safe default instead of exposing to config option to the user
-	// in order to simplify the config.
-	RingNumTokens = 512
-
 	// sharedOptionWithRingClient is a message appended to all config options that should be also
 	// set on the components running the store-gateway ring client.
 	sharedOptionWithRingClient = " This option needs be set both on the store-gateway, querier and ruler when running in microservices mode."
@@ -77,6 +72,7 @@ type RingConfig struct {
 	HeartbeatTimeout     time.Duration `yaml:"heartbeat_timeout" category:"advanced"`
 	ReplicationFactor    int           `yaml:"replication_factor" category:"advanced"`
 	TokensFilePath       string        `yaml:"tokens_file_path"`
+	NumTokens            int           `yaml:"num_tokens" category:"advanced"`
 	ZoneAwarenessEnabled bool          `yaml:"zone_awareness_enabled"`
 	AutoForgetEnabled    bool          `yaml:"auto_forget_enabled"`
 
@@ -115,6 +111,7 @@ func (cfg *RingConfig) RegisterFlags(f *flag.FlagSet, logger log.Logger) {
 	f.IntVar(&cfg.ReplicationFactor, ringFlagsPrefix+"replication-factor", 3, "The replication factor to use when sharding blocks."+sharedOptionWithRingClient)
 	f.StringVar(&cfg.TokensFilePath, ringFlagsPrefix+"tokens-file-path", "", "File path where tokens are stored. If empty, tokens are not stored at shutdown and restored at startup.")
 	f.BoolVar(&cfg.ZoneAwarenessEnabled, ringFlagsPrefix+"zone-awareness-enabled", false, "True to enable zone-awareness and replicate blocks across different availability zones."+sharedOptionWithRingClient)
+	f.IntVar(&cfg.NumTokens, ringFlagsPrefix+"num-tokens", 512, "Number of tokens for each store-gateway.")
 	f.BoolVar(&cfg.AutoForgetEnabled, ringFlagsPrefix+"auto-forget-enabled", true, fmt.Sprintf("When enabled, a store-gateway is automatically removed from the ring after failing to heartbeat the ring for a period longer than %d times the configured -%s.", ringAutoForgetUnhealthyPeriods, ringHeartbeatTimeoutFlag))
 
 	// Wait stability flags.
@@ -164,7 +161,7 @@ func (cfg *RingConfig) ToLifecyclerConfig(logger log.Logger) (ring.BasicLifecycl
 		HeartbeatPeriod:                 cfg.HeartbeatPeriod,
 		HeartbeatTimeout:                cfg.HeartbeatTimeout,
 		TokensObservePeriod:             0,
-		NumTokens:                       RingNumTokens,
+		NumTokens:                       cfg.NumTokens,
 		KeepInstanceInTheRingOnShutdown: !cfg.UnregisterOnShutdown,
 	}, nil
 }